fix(editor): symbolKind.valueSet 을 i32 로 — u8 배열이 JSON 문자열이 돼 tsgo 가 initialize 를 거부하고 세션이 통째로 죽었다; 빈 selectionRange 는 심볼 범위에서 이름을 찾아 쓰고, 피커는 1층 provider 를 전제하지 않는다 (§8.2o)

+++ src/session/lsp/rpc.zig
@@ -160,7 +160,9 @@ pub fn initializeRequest(allocator: std.mem.Allocator, root_uri: []const u8, pid
                     .documentSymbol = .{
                         .dynamicRegistration = false,
                         .hierarchicalDocumentSymbolSupport = true,
-                        .symbolKind = .{ .valueSet = [_]u8{ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26 } },
+                        // **`u8` 배열은 JSON 문자열이 된다**(Zig `Stringify` 의 규칙) — tsgo 가 `cannot unmarshal JSON string into []SymbolKind` 로
+                        // `initialize` 를 통째로 거부했고 세션 전체가 죽었다(2026-09-22 tee 실측 — `initialized` params 사고와 같은 부류). `i32` 로 낸다.
+                        .symbolKind = .{ .valueSet = [_]i32{ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26 } },
                     },
                     // 구현·타입 정의·선언(§8.2m) — 같은 피커. `linkSupport` 로 `LocationLink[]` 를 받는다(tsgo 가 그것을 낸다).
                     .implementation = .{ .dynamicRegistration = false, .linkSupport = true },
